スコアの取得と登録の実装(testなし) #13 #14

+++ pkg/models/dto.go
@@ -14,7 +14,7 @@ type (
 
 	Score struct {
 		UserId string `json:"userid"`
-		Score  int    `json:"score"`
+		Score  int64  `json:"score"`
 	}
 
 	UserId struct {
